+++ internal/collector/nodeexporter/memory.go
@@ -23,7 +23,7 @@ package nodeexporter
 import (
 	"fmt"
 
-	"github.com/shirou/gopsutil/v3/mem"
+	"github.com/shirou/gopsutil/v4/mem"
 
 	"github.com/telemetryflow/telemetryflow-agent/internal/collector"
 )